Fix widths (#24)
* Update badge widths

Switch to using get_text_width(' ') to get padding width.
This results in a more compact badge as the font_width that was used before
was the guesstimated font width that was too large.

* Update default num padding chars

This commit updates the number of padding characters to zero.  Many issues have been
raised around producing more compact badges, so it feels like the most common usage would be
with this option defaulted to zero.