test: Ensure tests run against correct version
Update the `inv --test.local` command to ensure that tests are run against
the correct version of the package. The invoke command has been updated to:

- Ensure the `anybadge` command is not available on the system to start with

- Ensure `pip` is available

- Pip install the local project

- Run tests

- Pip uninstall `anybadge`

The net result is that invoke will be more robust in terms of making sure the
installed version is in sync when running CLI tests.